feat: Renamed app folder to accounts"
This also includes addition of LookUps to the accounts app, moving CustomerUser from authentication to accounts
