+++ project/accounts/admin.py
