--- project/app/models.py
@@ -1,10 +1,0 @@
-# -*- encoding: utf-8 -*-
-"""
-Copyright (c) 2019 - present AppSeed.us
-"""
-
-from django.db import models
-from django.contrib.auth.models import User
-
-# Create your models here.
-
