+++ project/authentication/forms.py
@@ -6,7 +6,7 @@ Copyright (c) 2019 - present AppSeed.us
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
 # from django.contrib.auth.models import User
-from .models import CustomUser
+from accounts.models import CustomUser
 
 
 class LoginForm(forms.Form):
